requirements.txt
python-decouple==3.6
pytz==2022.1
pyxdg==0.27
django==4.1.3
djangorestframework==3.14.0
Pillow==9.3.0
asgiref==3.5.2
sqlparse==0.4.3

